+++ packages/react/src/components/Toast/index.tsx
@@ -1,6 +1,6 @@
 import * as ToastPrimitive from "@radix-ui/react-toast";
 import { X } from "phosphor-react";
-import { ReactNode } from "react";
+import { Dispatch, ReactNode, SetStateAction } from "react";
 
 import {
   ToastCloseButton,
@@ -13,11 +13,19 @@ import {
 interface ToastProviderProps {
   children: ReactNode;
   swipeDirection: "down" | "left" | "right" | "up";
+  duration: number;
 }
 
-function ToastProvider({ children, swipeDirection }: ToastProviderProps) {
+function ToastProvider({
+  children,
+  swipeDirection,
+  duration
+}: ToastProviderProps) {
   return (
-    <ToastPrimitive.Provider swipeDirection={swipeDirection} label="teste">
+    <ToastPrimitive.Provider
+      swipeDirection={swipeDirection}
+      duration={duration}
+    >
       {children}
     </ToastPrimitive.Provider>
   );
@@ -27,13 +35,20 @@ export interface ToastProps {
   title: string;
   description: string;
   open: boolean;
+  setOpen: Dispatch<SetStateAction<boolean>>;
   close: () => void;
 }
 
-export function ToastElement({ title, description, open, close }: ToastProps) {
+function ToastElement({
+  title,
+  description,
+  open,
+  setOpen,
+  close
+}: ToastProps) {
   return (
     <>
-      <ToastRoot open={open}>
+      <ToastRoot open={open} onOpenChange={setOpen}>
         <ToastTitle> {title} </ToastTitle>
         <ToastDescription>{description}</ToastDescription>