+++ packages/react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ignite-ui-caio/react
 
+## 5.0.0
+
+### Major Changes
+
+- Adjustes in Toast
+
+## 4.0.0
+
+### Major Changes
+
+- Adjustments in Toast Swipe close
+
+## 3.1.0
+
+### Minor Changes
+
+- Adjustments in Toast
+
+## 3.0.0
+
+### Major Changes
+
+- Add Toast component
+
 ## 2.0.0
 
 ### Major Changes